turn off vestingsc tests
